split files classes into their own files in /physics
