+++ physics_development/index.html
@@ -4,7 +4,7 @@
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Physics Development</title>
-	<script type="module" src="js/cookie.js"></script>
+	<script type="module" src="js/core.js"></script>
 	<style>
 		*, ::before, ::after {
 			box-sizing: border-box;
