+++ src/components/hero.astro
@@ -33,7 +33,7 @@ import { content } from "../content";
 	<div class="cookie-bits"></div>
 </section>
 
-<script src="../cookie.ts"></script>
+<script src="../physics/core.ts"></script>
 
 <style define:vars={{"bg-pattern": `url("${backgroundPatterns.whirl(colors.darker, .3)}")`}}>
 	section {
